refactor(test): J-Quants認証関連テストをシステムテストに移動
tests/unit/jquants/test_client.py に存在したJ-Quants認証関連のユニットテストを、`tests/system/jquants/test_auth.py` に移動した。

これにより、ユニットテストとシステムテストの責務を明確に分離し、テストコードの可読性とメンテナンス性を向上させた。

`test_client.py` は低レベルなクライアント機能のテストに集中し、`test_auth.py` は実際の認証フローと設定ファイル連携を含むシステムレベルのテストに集中する。
